refacto(FiltersModule): Rename connector to Difficulties
It was some legacy naming back when difficulties
was the only submodule in filters
