+++ frontend/src/modules/filters/difficulties/adapter.ts
