+++ frontend/src/modules/filters/difficulties/api.ts
