+++ frontend/src/modules/filters/difficulties/connector.ts
