+++ frontend/src/modules/filters/difficulties/index.ts
