+++ frontend/src/modules/filters/difficulties/interface.ts
